+++ TASVideos.Core/Extensions/WikiHelper.cs
@@ -1,4 +1,5 @@
 ﻿using System.Text.RegularExpressions;
+using TASVideos.Core.Services.Wiki;
 using TASVideos.Data.Entity;
 
 namespace TASVideos.Extensions;
