+++ TASVideos.Core/Services/Youtube/Dtos/IWikiToTextRenderer.cs
@@ -1,4 +1,4 @@
-﻿using TASVideos.Data.Entity;
+﻿using TASVideos.Core.Services.Wiki;
 
 namespace TASVideos.Core.Services.Youtube;
