+++ TASVideos/Pages/RssFeeds/Models/RssPublication.cs
@@ -1,4 +1,5 @@
-﻿using TASVideos.Data.Entity;
+﻿using TASVideos.Core.Services.Wiki;
+using TASVideos.Data.Entity;
 
 namespace TASVideos.Pages.RssFeeds.Models;
