+++ TASVideos/Pages/Shared/Components/GameResourcesFooter/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@model IWikiPage
+﻿@model TASVideos.Core.Services.Wiki.IWikiPage
 @await Component.ListSubPages(Model, true)
 <div role="alert" class="alert alert-info">
 	@await Component.RenderWiki("System/GameResourcesFooter")
