+++ TASVideos/Pages/Shared/_WikiLayoutFooter.cshtml
@@ -1,4 +1,6 @@
-﻿@model IWikiPage
+﻿@using TASVideos.TagHelpers
+@using Microsoft.AspNetCore.Mvc.TagHelpers
+@model TASVideos.Core.Services.Wiki.IWikiPage
 @{
 	bool canEdit = WikiHelper.UserCanEditWikiPage(Model.PageName, User.Name(), User.Permissions());
 }
