+++ TASVideos/Pages/Shared/_WikiLayoutHeader.cshtml
@@ -1,4 +1,5 @@
-﻿@model IWikiPage
+﻿@using Microsoft.AspNetCore.Mvc.TagHelpers
+@model TASVideos.Core.Services.Wiki.IWikiPage
 @{
 	bool canEdit = WikiHelper.UserCanEditWikiPage(Model.PageName, User.Name(), User.Permissions());
 }
