+++ TASVideos/TagHelpers/WikiMarkupTagHelper.cs
@@ -1,14 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Text.Encodings.Web;
-using System.Threading.Tasks;
+﻿using System.Text.Encodings.Web;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.AspNetCore.Mvc.ViewFeatures;
 using Microsoft.AspNetCore.Razor.TagHelpers;
-using TASVideos.Data.Entity;
-using TASVideos.Extensions;
+using TASVideos.Core.Services.Wiki;
 using TASVideos.Services;
 using TASVideos.WikiEngine;
 using TASVideos.WikiEngine.AST;
