+++ TASVideos/ViewComponents/Frames.cs
@@ -2,8 +2,8 @@
 using Microsoft.EntityFrameworkCore;
 using TASVideos.Common;
 using TASVideos.Core.Services;
+using TASVideos.Core.Services.Wiki;
 using TASVideos.Data;
-using TASVideos.Data.Entity;
 using TASVideos.WikiEngine;
 
 namespace TASVideos.ViewComponents;
