+++ TASVideos/ViewComponents/GameResourcesFooter.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using TASVideos.Data.Entity;
+using TASVideos.Core.Services.Wiki;
 
 namespace TASVideos.ViewComponents;
