+++ TASVideos/ViewComponents/ListLanguages.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ViewComponents;
 using TASVideos.Core.Services;
-using TASVideos.Data.Entity;
+using TASVideos.Core.Services.Wiki;
 
 namespace TASVideos.ViewComponents;
